Reflected comments for PR. moved get_samples to audio_utils
